Add infrastructure config modules and update aliases
